Fix compilation warning deteced on XCode (#993)

+++ samples/extensions/dynamic_blending/dynamic_blending.cpp
@@ -115,7 +115,7 @@ void DynamicBlending::prepare_scene()
 
 	index_count = static_cast<uint32_t>(indices.size());
 
-	vertex_buffer_size     = vertices.size() * sizeof(Vertex);
+	vertex_buffer_size     = static_cast<uint32_t>(vertices.size() * sizeof(Vertex));
 	auto index_buffer_size = indices.size() * sizeof(uint32_t);
 
 	vertex_buffer = std::make_unique<vkb::core::Buffer>(get_device(),
